app/web/requirements.txt
imgaug==0.4.0
joblib==1.3.2
keras==2.15.0
numpy==1.26.2
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==9.0.1
Pillow==10.1.0
streamlit==1.28.2
tensorflow==2.15.0
tensorflow_macos==2.15.0
